% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionPrep.R
\name{additionPrep}
\alias{additionPrep}
\title{Bind species to hypothesized relatives in a phylogeny}
\usage{
additionPrep(tree, addition.statements, missing.sp)
}
\arguments{
\item{tree}{An ape-style phylogenetic tree.}

\item{addition.statements}{A data frame of taxonomy, including species currently
in the tree, as well as all species that need to be added. The data frame needs
to contain the following columns: "species","family","genus","add.to", and 
"do.not.break". "add.to" now respects unlimited addition statements, each more specific
than the last, separated by a semi-colon. Families, genera, and species groups
are all permitted. Species groups need to have species separated by commas.
Large clades can be targeted by specifying species that span the desired MRCA.
The exclusion statements, contained in the "do.not.break" column, can also
take the form of families, genera, and species groups, and are also separated
by semi-colons. These are additive, i.e. two or more clades can be excluded at
once.}

\item{missing.sp}{The name of the missing species.}
}
\value{
A data frame specifying the node to which to bind the missing species,
as well as details about how that selection was made.
}
\description{
Given a data frame of species and taxonomic assignments, and a phylogeny with
some of those species in it, will determine the allowed positions for a
particular missing taxon.
}
\details{
The work horse behind customAdd(). See additional details in that
function.
}
\references{
Mast et al. 2015. Paraphyly changes understanding of timing and tempo of 
diversification in subtribe Hakeinae (Proteaceae), a giant Australian plant radiation.
American Journal of Botany.
}
