% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaxa.R
\name{addTaxa}
\alias{addTaxa}
\title{Bind species to hypothesized relatives in a phylogeny}
\usage{
addTaxa(tree, groupings, branch.position = "midpoint", ini.lambda = 1,
  ini.mu = 0, no.trees, clade.membership, crown.can.move = TRUE)
}
\arguments{
\item{tree}{An ape-style phylogenetic tree.}

\item{groupings}{A data frame with two columns, "species" and "group". Missing species,
to be added, are taken as those that do not match a value in the tip labels of tree.}

\item{branch.position}{Once the algorithm has selected to which node a species will
be bound, this determines where on the branch that happens. Note that this refers to
the position below (stemwards) from the node the tip will be bound.Currently there are
four options. The first, "polytomy", creates a polytomy at the node to which the new
tip is added (need to check but it might actually make a dichotomous branch of length 0).
The second, "midpoint", simply splits the difference between the node and its parent
and binds the new tip in there.
The third, "uniform", will sample from a uniform distribution with a minimum of
zero and a maximum of the full distance between the node and its parent. Note that
this means that species can conceivably be added at the maximum or minimum
distance possible from the tip, which would create a polytomy (or branch of length 0?).
I have removed checks to account and deal with this--if you need them back let me know. 
The fourth option is
"bd." This uses the corsim function from the TreeSim package to simulate the missing
speciation events according to speciation (lambda) and extinction (mu) values passed
to addTaxa.}

\item{ini.lambda}{Initial speciation value for the "bd" optimization, if that option
of branch position is chosen. Defaults to 1.}

\item{ini.mu}{Initial extinction value for the "bd" optimization, if that option
of branch position is chosen. Defaults to 0.1.}

\item{no.trees}{The number of desired final trees with all missing species from 
groupings added.}

\item{clade.membership}{An optional data frame with first column = "species", second
column = "clade". These are named, monophyletic clades to which the species in the
input phylogeny can belong. Not every species in the input phylogeny needs to be in the
clade membership data frame, but missing species added to species not in this frame
will not be included in the output clade membership data frame. Such species will
still be included in the output phylogenies. Note that these clades need to be
mutually exclusive. For example, clade 2 cannot be contained within clade 1.}

\item{crown.can.move}{Logical. If TRUE, and if missing taxa are to be added stemwards,
this will allow the age of the crown group to potentially shift back in time. If FALSE,
and if missing taxa were to be added stemwards, this will prevent taxa from being added
below the crown group. It will force these to be bound crownwards if the crown node
in the clade is selected to be bound to. Note that the argument
name is somewhat misleading in that the crown ages can still shift when missing taxa
are added. Specifically, if missing taxa are bound to what were initially
single-species clades in the input tree, then the
crown age will "shift" forward in time (the single-taxon clade did not actually have a 
crown age). If crown.can.move is set to FALSE, then after one taxon is added to such a
single-species clade, the crown age of that clade then becomes fixed and will not move.}
}
\value{
A list with two elements: (1) multiPhylo object with number of trees as 
determined by no.trees, and, if clade.membership was provided, (2) a list of data 
frames summarizing which named clade each species in the complete phylogeny belongs to
(if any).
}
\description{
Given a data frame of species and taxonomic assignments, and an accepted phylogeny with
some of those species in it, will add the missing species in next to a taxonomic 
relative. Optionally, can keep track of which named clades species are bound to.
}
\details{
REVISE! Given a data frame of two columns, "species" and "group", will take a species
that is absent from the phylogeny and bind it to a randomly selected taxonomic 
relative. Note that the current implementation of the function iteratively builds up
the tree, meaning that a species being bound into the tree can be bound to a
species that was in the input phylogeny, or to a species that was added during a
previous iteration of the function. Previously, added species could only be bound to
species that were in the input phylogeny. This new feature slows the function down,
but it results in more realistic-looking, less ladderized trees. Additionally, previous
iterations of this function also worked from top to bottom of the list of missing taxa.
This potentially results in a biased exploration of possible tree space. The function
now samples missing taxa and updates the missing list accordingly to allow a less
biased exploration of possible complete trees. Four distinct methods
of adding new taxa are possible. With "polytomy", the new species is simply assigned to
a tip. Both species end up with branch lengths to their most recent common ancestor of
0. With "crown", the new species is bound in at half the distance between the species
it is being bound to and that species' original parent node. With "stem", the new
species is bound to the parent node of the species it was selected to be bound to. The
new species is bound in at half the distance between the parent node and grandparent
node. With "randomly", each new species is randomly bound either crown-wards or
stem-wards, following the descriptions above. Note that even if "stem" or "randomly" 
are chosen, if a species is to be bound to the sister species to all others (the most
"basal" species in the phylogeny), it will automatically be bound stem-wards. With all
options, if the input tree is ultrametric, the output trees should remain so. 
Currently, no effort is made to ensure that the taxonomic groups of the missing species
are actually to be found in the species in the input tree.
}
\examples{
data(chelonia)
tree <- chelonia$phy

#some species in this tree are identified to subspecies. drop those
temp <- lapply(strsplit(tree$tip.label, "_"), length)
names(temp) <- tree$tip.label
temp <- temp[temp==2]
tree <- drop.tip(tree, setdiff(tree$tip.label, names(temp)))

#create an example groupings data frame.
groupsDF <- data.frame(species=tree$tip.label)
groupsDF$group <- unlist(lapply(strsplit(tree$tip.label, "_"), "[", 1))

#use the function to drop 100 species (there were 194 in tree)
example <- tipDropper(tree, groupsDF, 100)

#add those missing species back in
newTrees <- addTaxa(tree=example, groupings=groupsDF, branch.position="bd",
  no.trees=1)
}
\references{
Mast et al. 2015. Paraphyly changes understanding of timing and tempo of 
diversification in subtribe Hakeinae (Proteaceae), a giant Australian plant radiation.
American Journal of Botany.
}
