% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bdScaler.R
\name{bdScaler}
\alias{bdScaler}
\title{Generate a speciation event}
\usage{
bdScaler(tree, lambda, mu, min.age, max.age)
}
\arguments{
\item{tree}{A phylogeny in ape format.}

\item{lambda}{The speciation rate.}

\item{mu}{The exctinction rate.}

\item{min.age}{Minimum age of the missing speciation event to be inferred. Note that
this is the age counting back from the tips, such that 0 is the present. Thus, min.age
must always be less than max.age.}

\item{max.age}{Maximum age of the missing speciation event to be inferred.}
}
\value{
A missing speciation time.
}
\description{
Place speciation event in a speciation/extinction rate-informed position
}
\details{
This function assumes that a single species is missing, then takes the
provided lambda and mu values, and the range of time over which the speciation event
occurred, and generates the missing time.
}
\examples{
#load a molecular tree up
data(bird.families)

bdScaler(bird.families, lambda=1, mu=0, min.age=0, max.age=10)
}
\references{
ETM unpublished
}

