% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/branchScaler.R
\name{branchScaler}
\alias{branchScaler}
\title{Scale a vector of branch lengths}
\usage{
branchScaler(input.vector, max.age)
}
\arguments{
\item{input.vector}{A vector of numbers such as the length of all the branches in
a phylogeny.}

\item{max.age}{The maximum value you want to occur in the vector.}
}
\value{
A scaled vector.
}
\description{
Simple utility function to scale tree-wide distribution of branch lengths
}
\details{
This function simply divides the input vector, e.g. a distribution of all the
edge lengths in a tree, by the maximum length in that vector, then multiplies the
resulting values by the "max.age" argument.
}
\examples{
#load a molecular tree up
data(bird.families)

branchScaler(bird.families$edge.length, 2)
}
\references{
Mast et al. 2015. Paraphyly changes understanding of timing and tempo of
diversification in subtribe Hakeinae (Proteaceae), a giant Australian plant radiation.
American Journal of Botany.
}

