% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divRates.R
\name{divRates}
\alias{divRates}
\title{Calculate sensitivity of absolute diversification rates to missing taxa}
\usage{
divRates(tree, groupings, branch.position, ini.lambda = 1, ini.mu = 0.1,
  no.trees, clade.membership, crown.can.move, calc.from.crown, epsilon)
}
\arguments{
\item{tree}{An ape-style phylogenetic tree}

\item{groupings}{A data frame with two columns, "species" and "group". Missing species,
to be added, are taken as those that do not match a value in the tip labels of tree.}

\item{branch.position}{Once the algorithm has selected to which node a species will
be bound, this determines where on the branch that happens. Note that this refers to
the position below (stemwards) from the node the tip will be bound.Currently there are
four options. The first, "polytomy", creates a polytomy at the node to which the new
tip is added (need to check but it might actually make a dichotomous branch of length 0).
The second, "midpoint", simply splits the difference between the node and its parent
and binds the new tip in there.
The third, "uniform", will sample from a uniform distribution with a minimum of
zero and a maximum of the full distance between the node and its parent. Note that
this means that species can conceivably be added at the maximum or minimum
distance possible from the tip, which would create a polytomy (or branch of length 0?).
I have removed checks to account and deal with this--if you need them back let me know. 
The fourth option is
"bd." This uses the corsim function from the TreeSim package to simulate the missing
speciation events according to speciation (lambda) and extinction (mu) values passed
to addTaxa.}

\item{ini.lambda}{Initial speciation value for the "bd" optimization, if that option
of branch position is chosen. Defaults to 1.}

\item{ini.mu}{Initial extinction value for the "bd" optimization, if that option
of branch position is chosen. Defaults to 0.1.}

\item{no.trees}{The number of desired final trees with all missing species from 
groupings added.}

\item{clade.membership}{An optional data frame with first column = "species", second
column = "clade". These are named, monophyletic clades to which the species in the
input phylogeny can belong. Not every species in the input phylogeny needs to be in the
clade membership data frame, but missing species added to species not in this frame
will not be included in the output clade membership data frame. Such species will
still be included in the output phylogenies. Note that these clades need to be
mutually exclusive. For example, clade 2 cannot be contained within clade 1.}

\item{crown.can.move}{Logical. If TRUE, and if missing taxa are to be added stemwards,
this will allow the age of the crown group to potentially shift back in time. If FALSE,
and if missing taxa were to be added stemwards, this will prevent taxa from being added
below the crown group. It will force these to be bound crownwards if the crown node
in the clade is selected to be bound to. Note that the argument
name is somewhat misleading in that the crown ages can still shift when missing taxa
are added. Specifically, if missing taxa are bound to what were initially
single-species clades in the input tree, then the
crown age will "shift" forward in time (the single-taxon clade did not actually have a 
crown age). If crown.can.move is set to FALSE, then after one taxon is added to such a
single-species clade, the crown age of that clade then becomes fixed and will not move.}

\item{calc.from.crown}{Whether the crown or stem method of bd.ms from geiger will be
used.}

\item{epsilon}{The extinction rate parameter for bd.ms}
}
\value{
A dataframe where each row corresponds to a complete tree, and each column
corresponds to the calculated absolute diversification rate of a named clade.
}
\description{
Given a data frame of species and taxonomic assignments, an accepted phylogeny with
some of those species in it, and a number of named, monophyletic clades, will add
missing species next to a taxonomic relatives and quantify sensitivity of
clade-specific absolute diversification rates.
}
\details{
Adds missing taxa according to all rules and details outlined in
addTaxa(). During this process, it tabulates to which named clades missing
species are added, and calculates the diversification rate of each clade after each
complete tree is created. In this way, the sensitivity of these absolute
diversification rates to missing taxa can be assessed.
}
\examples{
data(chelonia)
tree <- chelonia$phy

#some species in this tree are identified to subspecies. drop those
temp <- lapply(strsplit(tree$tip.label, "_"), length)
names(temp) <- tree$tip.label
temp <- temp[temp==2]
tree <- drop.tip(tree, setdiff(tree$tip.label, names(temp)))

#create an example groupings data frame.
groupsDF <- data.frame(species=tree$tip.label)
groupsDF$group <- unlist(lapply(strsplit(tree$tip.label, "_"), "[", 1))

#and create a dummy data frame to pay attention to diversification rates
#of two arbitrary clades
graptemysEtc <- data.frame(
  species=groupsDF$species[groupsDF$group \%in\% c("Graptemys","Trachemys","Malaclemys")],
  clade="graptemys")
gopherus <- data.frame(species=tree$tip.label[grep("Gopherus", tree$tip.label)],
  clade="gopherus")
cladesDF <- rbind(graptemysEtc, gopherus)

#drop species, add them back in and see how diversification rates of those clades
#change
example <- tipDropper(tree, groupsDF, 10)

sensitivity <- divRates(tree=example, groupings=groupsDF, branch.position="bd",
  no.trees=10, clade.membership=cladesDF, crown.can.move=TRUE,
  calc.from.crown=TRUE, epsilon=0.1)
}
\references{
Mast et al. 2015. Paraphyly changes understanding of timing and tempo of 
diversification in subtribe Hakeinae (Proteaceae), a giant Australian plant radiation.
American Journal of Botany.
}
