% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isSingleton}
\alias{isSingleton}
\title{Determine whether a node leads to a tip}
\usage{
isSingleton(tree, nodes, crown.can.move)
}
\arguments{
\item{tree}{Phylogeny in ape-format}

\item{nodes}{Vector of node numbers, with an optional node named root.}

\item{crown.can.move}{True or false. Related to whether you want to allow missing taxa
to be added stemwards from the input crown data of the clade in question. See details.}
}
\value{
A vector of TRUE/FALSE corresponding to the rules detailed above.
}
\description{
Takes a vector of node numbers and determines whether it leads to a tip or to an
internal node (plus more).
}
\details{
This function determines what descends from a node. In the context of the
package, these will generally be nodes that subtend named clades. It takes a vector of
node numbers, with an optional node named "root". If no such root is provided,
then it will not evaluate whether there would be problems with binding taxa to the
species that is sister to all others (if one exists). If a tip is provided it
returns TRUE. If an internal node that leads to other internal nodes is provided, it
returns FALSE. This is because any such node will always have taxa bound crownwards
from it; missing taxa are bound to species, not nodes, so adding stemwards to a species
descending from this node will mean the species is still crownwards from the node in
question. Finally, if an internal node is provided that leads to at least one tip,
the function returns TRUE if crown.can.move=FALSE, else returns FALSE.
}
\examples{
#load a molecular tree up
data(bird.families)

#define a named vector of node numbers, including a root
temp <- c()
temp[1] <- length(bird.families$tip.label) + 1
names(temp) <- "root"

#create a data frame of all taxa from the phylogeny, and make up clade memberships
#for each. note that the names on this data frame differ from "groupings" in other
#functions
dummy.frame <- data.frame(species=bird.families$tip.label,
clade=c(rep("nonPasserine", 95), rep("suboscine", 9), rep("basalOscine", 13),
rep("oscine", 20)))

#use the function getMRCAs() to determine the nodes subtending these named clades
nodes <- getMRCAs(bird.families, dummy.frame)
#unlist the results and append to the root node defined above
nodes <- append(temp, nodes)

#note effect of crown.can.move argument. because Acanthisittidae is sister
#to rest of suboscines, missing taxa cannot be bound stemwards from this tip or it will
#change the age of the suboscine clade.
isSingleton(bird.families, nodes, crown.can.move=TRUE)
isSingleton(bird.families, nodes, crown.can.move=FALSE)
}
\references{
Eliot Miller unpublished
}

