% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomlyAddTaxaOld.R
\name{randomlyAddTaxaOld}
\alias{randomlyAddTaxaOld}
\title{Bind species to hypothesized relatives in a phylogeny}
\usage{
randomlyAddTaxaOld(tree, groupings, from.node = "randomly",
  branch.position = "midpoint", optional.offset = 0, lambda = 1, mu = 0,
  no.trees, clade.membership, crown.can.move = TRUE,
  print.to.screen = FALSE, file.name)
}
\arguments{
\item{tree}{An ape-style phylogenetic tree}

\item{groupings}{A data frame with two columns, "species" and "group". Missing species,
to be added, are taken as those that do not match a value in the tip labels of tree.}

\item{from.node}{Whether species should be added in a "polytomy" with, "crown" (more 
recently diverged), "stem" (previously diverged), or "randomly" crown-wards or
stem-wards from the tip they are bound to.}

\item{branch.position}{Once the algorithm has selected to which branch a species will
be bound, this determines where on the branch that happens. Currently there are four
options. The first, "midpoint", simply breaks the branch in half and binds the species
there. The second, "uniform", will sample from a uniform distribution with a minimum of
zero and a maximum of the full branch length. The third, "overall", is probably not a
particularly good idea to use. It takes the distribution of branches from the full tree
and scales it to the branch in question, then samples from that. Thus, if there are
lots of long branches on average in the phylogeny, the species
being bound will tend to be bound closer to the subtending node. Since this is just
based on the tree-wide distribution of branch lengths as opposed to the lengths in that
"portion" of the tree, this is not currently recommended. Note that options 2 and 3
(particularly "overall") mean that species can conceivably be added at the maximum 
distance possible from the tip, i.e. at the node subtending them. This would create
lengths of zero. To avoid this, the argument "optional.offset" can be set to some small
value, e.g. 1e-9. This argument does not currently affect uniform, thus there is a very
small chance that a polytomy could be created with that option. The final option is
"bd." This uses the corsim function from the TreeSim package to simulate the missing
speciation events according to speciation (lambda) and extinction (mu) values passed
to randomlyAddTaxa. This approach is currently in testing.}

\item{optional.offset}{Value to avoid polytomies (see above).}

\item{lambda}{The speciation rate, such as that calculated with diversitree or TreePar.}

\item{mu}{The extinction rate, such as that calculated with diversitree or TreePar.}

\item{no.trees}{The number of desired final trees with all missing species from 
groupings added.}

\item{clade.membership}{An optional data frame with first column = "species", second
column = "clade". These are named, monophyletic clades to which the species in the
input phylogeny can belong. Not every species in the input phylogeny needs to be in the
clade membership data frame, but missing species added to species not in this frame
will not be included in the output clade membership data frame. They will of course
still be included in the output phylogenies.}

\item{crown.can.move}{Logical. If TRUE, and if missing taxa are to be added stemwards,
this will allow the age of the crown group to potentially shift back in time. If FALSE,
and if missing taxa were to be added stemwards, this will prevent taxa from being added
below the crown group. It will force these to be bound crownwards if the "basal"
taxon in the clade is selected as the species to be bound to. Note that the argument
name is somewhat misleading in that the crown ages can still shift when missing taxa
are added. Specifically, if missing taxa are bound to single-species clades, then the
crown age will "shift" forward in time (the single-taxon clade did not actually have a 
crown age).}

\item{print.to.screen}{Logical. Will print a list of the missing taxa to screen.}

\item{file.name}{Optional name for output RData file. If not provided then the object
will be stored in memory instead of saved to workspace. Either way, it is calculated in
memory. The function could probably be sped up by having this save directly to file,
appending each tree instead of first saving them all to memory. The saved object can be
loaded with load(), and will be imported with the name "taxaAddResults".}
}
\value{
A list with two elements: (1) multiPhylo object with number of trees as 
determined by no.trees, and, if clade.membership was provided, (2) a list of data 
frames summarizing which named clade each species in the complete phylogeny belongs to
(if any).
}
\description{
Given a data frame of species and taxonomic assignments, and an accepted phylogeny with
some of those species in it, will add the missing species in next to a taxonomic 
relative. Optionally, can keep track of which named clades species are bound to.
}
\details{
Given a data frame of two columns, "species" and "group", will take a species
that is absent from the phylogeny and bind it to a randomly selected taxonomic 
relative. Note that the current implementation of the function iteratively builds up
the tree, meaning that a species being bound into the tree can be bound to a
species that was in the input phylogeny, or to a species that was added during a
previous iteration of the function. Previously, added species could only be bound to
species that were in the input phylogeny. This new feature slows the function down,
but it results in more realistic-looking, less ladderized trees. Additionally, previous
iterations of this function also worked from top to bottom of the list of missing taxa.
This potentially results in a biased exploration of possible tree space. The function
now samples missing taxa and updates the missing list accordingly to allow a less
biased exploration of possible complete trees. Four distinct methods
of adding new taxa are possible. With "polytomy", the new species is simply assigned to
a tip. Both species end up with branch lengths to their most recent common ancestor of
0. With "crown", the new species is bound in at half the distance between the species
it is being bound to and that species' original parent node. With "stem", the new
species is bound to the parent node of the species it was selected to be bound to. The
new species is bound in at half the distance between the parent node and grandparent
node. With "randomly", each new species is randomly bound either crown-wards or
stem-wards, following the descriptions above. Note that even if "stem" or "randomly" 
are chosen, if a species is to be bound to the sister species to all others (the most
"basal" species in the phylogeny), it will automatically be bound stem-wards. With all
options, if the input tree is ultrametric, the output trees should remain so. 
Currently, no effort is made to ensure that the taxonomic groups of the missing species
are actually to be found in the species in the input tree.
}
\examples{
#load a molecular tree up. resolve polytomies
data(bird.families)
bird.families <- multi2di(bird.families)

#create a data frame of all taxa from the phylogeny, and make up species groups
#for each.
dummyGroups <- data.frame(species=bird.families$tip.label, 
group=c(rep("nonPasserine", 95), rep("suboscine", 9), rep("basalOscine", 13), 
rep("oscine", 20)))

#now make up a few passerine taxa that are missing and add these into the dummy frame
toAdd <- data.frame(species=c("Icteria", "Yuhina", "Pityriasis", "Macgregoria"), 
group=c(rep("oscine", 2), rep("basalOscine", 2)))
groupsDF <- rbind(dummyGroups, toAdd)

#these groups were actually monophyletic. but make a slightly more detailed clade
#membership frame to see how one would be used. note that it doesn't include the
#missing taxa, and taxonomy does not exactly follow modern understanding
cladesDF <- data.frame(species=bird.families$tip.label, 
clade=c(rep("nonPasserine", 95), rep("suboscine", 9), rep("basalOscineOther", 6),
rep("basalOscineCore", 7), rep("oscineBase", 14), rep("oscineDerived", 6)))

#examples of changing the from.node argument. you can plot or write these trees out to
#better see what the differences are
crown <- randomlyAddTaxa(tree=bird.families, groupings=groupsDF, from.node="crown", 
	branch.position="midpoint", no.trees=10, clade.membership=cladesDF, 
crown.can.move=TRUE, print.to.screen=FALSE)
stem <- randomlyAddTaxa(tree=bird.families, groupings=groupsDF, from.node="stem", 
	branch.position="midpoint", no.trees=10, clade.membership=cladesDF, 
crown.can.move=TRUE, print.to.screen=FALSE)
polytomy <- randomlyAddTaxa(tree=bird.families, groupings=groupsDF,
from.node="polytomy", branch.position="midpoint", no.trees=10, 
clade.membership=cladesDF, crown.can.move=TRUE, print.to.screen=FALSE)
randomly <- randomlyAddTaxa(tree=bird.families, groupings=groupsDF,
from.node="randomly", branch.position="midpoint", no.trees=10, 
clade.membership=cladesDF, crown.can.move=TRUE, print.to.screen=FALSE)
}
\references{
Mast et al. 2015. Paraphyly changes understanding of timing and tempo of 
diversification in subtribe Hakeinae (Proteaceae), a giant Australian plant radiation.
American Journal of Botany.
}
