% Generated by roxygen2 (4.0.2): do not edit by hand
\name{singleSpeciesTree}
\alias{singleSpeciesTree}
\title{Create a single-species tree}
\usage{
singleSpeciesTree(tree, tip)
}
\arguments{
\item{tree}{An ape-style phylogenetic tree.}

\item{tip}{A character vector corresponding to one of the species in the phylogeny.}
}
\value{
A single-species phylo object.
}
\description{
Generate a single-species tree with appropriate stem branch length
}
\details{
Ape cannot readily deal with single-species trees. This functions will extract
a single-species tree with appropriate branch length to the most recent common
ancestor. The tree cannot be plotted, but it can be used for analyses.
}
\examples{
#load a molecular tree up
data(bird.families)

#extract a single-species tree
tree <- singleSpeciesTree(bird.families, "Rheidae")
}
\references{
Eliot Miller unpublished
}

