% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyMissing.R
\name{identifyMissing}
\alias{identifyMissing}
\title{Identify taxa that are not included in phylogeny}
\usage{
identifyMissing(tree, groupings)
}
\arguments{
\item{tree}{An ape-style phylogenetic tree}

\item{groupings}{A data frame with two columns, "species" and "group". Missing species,
to be added, are taken as those that do not match a value in the tip labels of tree.}
}
\value{
A data frame of missing taxa with two columns, species and group.
}
\description{
Given a data frame of species and taxonomic assignments, and an accepted phylogeny with
some of those species in it, identify which species are not included in the phylogeny.
}
\details{
Utility function to identify taxa in the groupings data frame that are not
included in the input phylogeny.
}
\examples{
#load a molecular tree up
data(bird.families)

#create a data frame of all taxa from the phylogeny, and make up species groups
#for each.
dummy.frame <- data.frame(species=bird.families$tip.label, 
group=c(rep("nonPasserine", 95), rep("suboscine", 9), rep("basalOscine", 13), 
rep("oscine", 20)))

#now make up a few passerine taxa that are missing and add these into the dummy frame
toAdd <- data.frame(species=c("Icteria", "Yuhina", "Pityriasis", "Macgregoria"), 
group=c(rep("oscine", 2), rep("basalOscine", 2)))
groupsDF <- rbind(dummy.frame, toAdd)

#use the function to see what was missing from the original tree as compared with the 
#new data frame
identifyMissing(bird.families, groupsDF)
}
\references{
Mast et al. 2015. Paraphyly changes understanding of timing and tempo of 
diversification in subtribe Hakeinae (Proteaceae), a giant Australian plant radiation.
American Journal of Botany.
}
